Move active Entity predicate to `Entities`.
