+++ gcd/src/main/java/org/spine3/server/storage/datastore/DsAggregateStorage.java
@@ -49,11 +49,11 @@ import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Predicates.not;
 import static org.spine3.base.Identifiers.idToString;
 import static org.spine3.server.aggregate.storage.AggregateField.aggregate_id;
-import static org.spine3.server.storage.datastore.DsProperties.activeEntityPredicate;
 import static org.spine3.server.storage.datastore.DsProperties.addArchivedProperty;
 import static org.spine3.server.storage.datastore.DsProperties.addDeletedProperty;
 import static org.spine3.server.storage.datastore.DsProperties.isArchived;
 import static org.spine3.server.storage.datastore.DsProperties.isDeleted;
+import static org.spine3.server.storage.datastore.Entities.activeEntity;
 
 /**
  * A storage of aggregate root events and snapshots based on Google Cloud Datastore.
@@ -164,7 +164,7 @@ public class DsAggregateStorage<I> extends AggregateStorage<I> {
 
         final Collection<Entity> aggregateEntityStates = Collections2.filter(
                 getEntityStates(),
-                not(activeEntityPredicate()));
+                not(activeEntity()));
         final Collection<Key> inactiveAggregateKeys = Collections2.transform(
                 aggregateEntityStates,
                 new Function<Entity, Key>() {
